12/5

+++ Level_1/SortAlg.h
@@ -8,18 +8,19 @@
 #include<time.h>
 #include"Queue.h"
 
-void SelectionSort(int *arr, int n, int &ss, int &gan);
-void InsertionSort(int *arr, int n, int &ss, int &gan);
-void InterchangeSort(int *arr, int n, int &ss, int &gan);
-void BubbleSort(int *arr, int n, int &ss, int &gan);
-void ShellSort(int *arr, int n, int &ss, int &gan);
-void QuickSort(int *arr, int l, int r, int &ss, int &gan);
-void MergeSort(int *arr, int l, int r, int &ss, int &gan);
-void RadixSort(int *arr, int n, int &ss, int &gan);
-void HeapSort(int *arr, int n, int &ss, int &gan);
-void ShakerSort(int *arr, int n, int &ss, int &gan);
+// Source_SortAlg.cpp
+void SelectionSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void InsertionSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void InterchangeSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void BubbleSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void ShellSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void QuickSort(int *arr, int l, int r, __int64 &ss, __int64 &gan);
+void MergeSort(int *arr, int l, int r, __int64 &ss, __int64 &gan);
+void RadixSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void HeapSort(int *arr, int n, __int64 &ss, __int64 &gan);
+void ShakerSort(int *arr, int n, __int64 &ss, __int64 &gan);
 
-
+// Source_SortCmp.cpp
 void SelectionSort_Report(int *arr, int n);
 void InterchangeSort_Report(int *arr, int n);
 void InsertionSort_Report(int *arr, int n);
